# specs/002-domain-model/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ODS Backend API — модель данных MVP
  description: |
    Канонический REST API backend (`002-domain-model`).
    Потребитель: `specs/003-portal-mvp` (зеркало в api-consumer.yaml).
    Граф проекта (`006-project-graph`): paths/components ниже; исходник — `specs/006-project-graph/contracts/openapi-graph.yaml`.
  version: 1.3.0
servers:
  - url: http://localhost:3000/api/v1
    description: Локальная разработка backend
  - url: /api/v1
    description: За reverse-proxy (frontend/nginx)
paths:
...
  /projects/{projectId}/analysis/language-report/latest:
    get:
      summary: Последний отчёт Language Detector
      operationId: getLatestLanguageReport
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LanguageReport'
        '404':
          description: Отчёт ещё не создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    LanguageEntry:
      type: object
      required: [language, file_count, sample_paths, parser_status]
      properties:
        language:
          type: string
        file_count:
          type: integer
        sample_paths:
          type: array
          items:
            type: string
        parser_id:
          type: string
          nullable: true
        parser_status:
          type: string
          enum: [available, missing, failed]
    LanguageReport:
      type: object
      required: [id, project_id, detected_at, languages]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        detected_at:
          type: string
          format: date-time
        languages:
          type: array
          items:
            $ref: '#/components/schemas/LanguageEntry'
...